grant: add named generic parameter for grant counts
Since we want the grant to automatically track upcalls, read-only
allows, and read-write allows, we need to specified all of those counts
on the generic type. Having bare number makes it really easy for a
developer to mix up the counts, so create a ZST wrapper that is type
safe that names the count at the call site.

In subsequent CL in this PR, the AllowRw and AllowRo counts will be
updated to non-zero number when each capsule is updated to use the new
interfaces (and drop the allow_readonly and allow_readwrite
SyscallDriver function).

Change-Id: I01af5b8af8169af1d93320f47a06b9880d41bcc0